Entering UserChoice into DB from Controller -- issue with generating ID WIP

+++ src/main/java/com/atspickem/backend/models/User.java
@@ -79,7 +79,7 @@ public class User {
         this.score = score;
     }
 
-    public float getPickRatio() {
+    public double getPickRatio() {
         return pickRatio;
     }
